# src/repositories/treatment-options/dto/update-treatment-option.dto.ts
import { PartialType } from '@nestjs/swagger';
import { CreateTreatmentOptionDto } from './create-treatment-option.dto';

export class UpdateTreatmentOptionDto extends PartialType(
  CreateTreatmentOptionDto,
) {}