# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/treatments/{id}": {
      "get": {
        "operationId": "TreatmentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del tratamiento",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tratamiento encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreatmentResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tratamiento no encontrado"
          }
        },
        "summary": "Obtener un tratamiento por ID",
        "tags": [
          "Treatments"
        ]
      },
      "patch": {
        "operationId": "TreatmentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del tratamiento a actualizar",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Datos a actualizar",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTreatmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tratamiento actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreatmentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos"
          },
          "404": {
            "description": "Tratamiento no encontrado"
          }
        },
        "summary": "Actualizar un tratamiento",
        "tags": [
          "Treatments"
        ]
      },
      "delete": {
        "operationId": "TreatmentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del tratamiento a eliminar",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tratamiento eliminado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Tratamiento no encontrado"
          }
        },
        "summary": "Eliminar un tratamiento (soft delete)",
        "tags": [
          "Treatments"
        ]
      }
    },
...
    "/treatment-options": {
      "post": {
        "operationId": "TreatmentOptionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTreatmentOptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Opción creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreatmentOptionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos"
          }
        },
        "summary": "Crear una nueva opción de tratamiento",
        "tags": [
          "treatment-options"
        ]
      },
      "get": {
        "operationId": "TreatmentOptionController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista paginada de opciones",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TreatmentOptionResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Obtener todas las opciones de tratamiento paginadas",
        "tags": [
          "treatment-options"
        ]
      }
    },
    "/treatment-options/paginated": {
      "get": {
        "operationId": "TreatmentOptionController_findPaginated",
        "parameters": [
          {
            "name": "term",
            "required": false,
            "in": "query",
            "description": "Término de búsqueda opcional.",
            "schema": {
              "example": "ejemplo",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número de página para la paginación (base 1).",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "description": "Número de elementos por página.",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Campo por el cual ordenar los resultados.",
            "schema": {
              "example": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Dirección del ordenamiento (ASC o DESC).",
            "schema": {
              "example": "DESC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de opciones",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationDto"
                }
              }
            }
          }
        },
        "summary": "Obtener todas las opciones de tratamiento paginadas",
        "tags": [
          "treatment-options"
        ]
      }
    },
    "/treatment-options/{id}": {
      "get": {
        "operationId": "TreatmentOptionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Opción encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreatmentOptionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Opción no encontrada"
          }
        },
        "summary": "Obtener una opción de tratamiento por ID",
        "tags": [
          "treatment-options"
        ]
      },
      "put": {
        "operationId": "TreatmentOptionController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTreatmentOptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Opción actualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreatmentOptionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Opción no encontrada"
          }
        },
        "summary": "Actualizar una opción de tratamiento existente",
        "tags": [
          "treatment-options"
        ]
      },
      "delete": {
        "operationId": "TreatmentOptionController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Opción marcada como eliminada"
          },
          "404": {
            "description": "Opción no encontrada"
          }
        },
        "summary": "Eliminar lógicamente una opción de tratamiento",
        "tags": [
          "treatment-options"
        ]
      }
    },
    "/treatment-options/type/{type}": {
      "get": {
        "operationId": "TreatmentOptionController_findByType",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de opciones del tipo especificado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TreatmentOptionResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Obtener opciones por tipo (vertical/horizontal)",
        "tags": [
          "treatment-options"
        ]
      }
    },
    "/treatment-options/best-match": {
      "get": {
        "operationId": "TreatmentOptionController_findBestOption",
        "parameters": [
          {
            "name": "heat",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "volume",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mejor opción encontrada o null",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreatmentOptionResponseDto"
                }
              }
            }
          }
        },
        "summary": "Encontrar la mejor opción basada en parámetros técnicos",
        "tags": [
          "treatment-options"
        ]
      }
    },
...
  },
  "info": {
    "title": "TermoLab API",
    "description": "API para gestión de tratamientos térmicos",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Identificador único de la entidad.",
            "example": 1
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha y hora de creación del registro.",
            "example": "2023-10-27T10:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha y hora de la última actualización del registro.",
            "example": "2023-10-27T10:30:00Z"
          },
          "status": {
            "type": "boolean",
            "description": "Estado del registro (activo/inactivo).",
            "example": true
          },
          "deleted": {
            "type": "boolean",
            "description": "Indicador de borrado lógico.",
            "example": false
          },
          "name": {
            "type": "string",
            "description": "Nombre completo del usuario",
            "example": "María González"
          },
          "birthDate": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha de nacimiento",
            "example": "1985-05-15"
          },
          "role": {
            "type": "string",
            "description": "Rol del usuario",
            "enum": [
              "admin",
              "operator"
            ],
            "example": "operator"
          },
          "email": {
            "type": "string",
            "description": "Correo electrónico único",
            "example": "maria.gonzalez@example.com"
          },
          "lastLogin": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha del último inicio de sesión",
            "example": "2023-10-30T14:30:00Z"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "name",
          "birthDate",
          "role",
          "email"
        ]
      },
      "PaginationDataDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "El número total de elementos en la colección completa.",
            "example": 100
          },
          "currentPage": {
            "type": "number",
            "description": "El número de la página actual (base 1).",
            "example": 1
          },
          "lastPage": {
            "type": "number",
            "description": "El número de la última página disponible.",
            "example": 10
          },
          "pageSize": {
            "type": "number",
            "description": "El número de elementos por página.",
            "example": 10
          },
          "totalPages": {
            "type": "number",
            "description": "El número total de páginas disponibles.",
            "example": 10
          },
          "hasNextPage": {
            "type": "boolean",
            "description": "Indica si hay una página siguiente disponible.",
            "example": true
          },
          "hasPrevPage": {
            "type": "boolean",
            "description": "Indica si hay una página anterior disponible.",
            "example": false
          }
        },
        "required": [
          "total",
          "currentPage",
          "lastPage",
          "pageSize",
          "totalPages",
          "hasNextPage",
          "hasPrevPage"
        ]
      },
      "PaginationDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "description": "Array de elementos para la página actual.",
            "items": {}
          },
          "paginationData": {
            "description": "Metadatos de paginación.",
            "allOf": [
              {
                "$ref": "#/components/schemas/PaginationDataDto"
              }
            ]
          }
        },
        "required": [
          "data",
          "paginationData"
        ]
      },
...
      "TreatmentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Identificador único de la entidad.",
            "example": 1
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha y hora de creación del registro.",
            "example": "2023-10-27T10:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha y hora de la última actualización del registro.",
            "example": "2023-10-27T10:30:00Z"
          },
          "status": {
            "type": "boolean",
            "description": "Estado del registro (activo/inactivo).",
            "example": true
          },
          "deleted": {
            "type": "boolean",
            "description": "Indicador de borrado lógico.",
            "example": false
          },
          "name": {
            "type": "string",
            "description": "Nombre del tratamiento",
            "example": "Tratamiento para crudo 18°API"
          },
          "description": {
            "type": "string",
            "description": "Descripción del tratamiento",
            "example": "Diseño para 500 bbl/día con 20% agua"
          },
          "type": {
            "type": "string",
            "description": "Tipo de tratador (vertical/horizontal)",
            "example": "horizontal",
            "enum": [
              "vertical",
              "horizontal"
            ]
          },
          "totalFlow": {
            "type": "number",
            "description": "Flujo total de emulsión (bpd)",
            "example": 500
          },
          "waterFraction": {
            "type": "number",
            "description": "Fracción de agua (%)",
            "example": 20
          },
          "inputTemperature": {
            "type": "number",
            "description": "Temperatura de entrada (°F)",
            "example": 75
          },
          "treatmentTemperature": {
            "type": "number",
            "description": "Temperatura de tratamiento (°F)",
            "example": 140
          },
          "ambientTemperature": {
            "type": "number",
            "description": "Temperatura ambiente (°F)",
            "example": 30
          },
          "oilRetentionTime": {
            "type": "number",
            "description": "Tiempo retención petróleo (min)",
            "example": 60
          },
          "waterRetentionTime": {
            "type": "number",
            "description": "Tiempo retención agua (min)",
            "example": 30
          },
          "windSpeed": {
            "type": "number",
            "description": "Velocidad del viento (mph)",
            "example": 15
          },
          "apiGravity": {
            "type": "number",
            "description": "Gravedad API del crudo",
            "example": 18
          },
          "calculatedOilFlow": {
            "type": "number",
            "description": "Flujo de petróleo calculado (bpd)",
            "example": 400
          },
          "calculatedWaterFlow": {
            "type": "number",
            "description": "Flujo de agua calculado (bpd)",
            "example": 100
          },
          "oilRetentionVolume": {
            "type": "number",
            "description": "Volumen retención petróleo (bbl)",
            "example": 16.67
          },
          "waterRetentionVolume": {
            "type": "number",
            "description": "Volumen retención agua (bbl)",
            "example": 2.08
          },
          "requiredHeat": {
            "type": "number",
            "description": "Calor requerido (BTU/hr)",
            "example": 262210
          },
          "heatLoss": {
            "type": "number",
            "description": "Pérdidas de calor (BTU/hr)",
            "example": 87120
          },
          "totalHeat": {
            "type": "number",
            "description": "Calor total (BTU/hr)",
            "example": 349330
          },
          "selectedDiameter": {
            "type": "number",
            "description": "Diámetro seleccionado (ft)",
            "example": 4
          },
          "selectedLength": {
            "type": "number",
            "description": "Longitud seleccionada (ft)",
            "example": 15
          },
          "designPressure": {
            "type": "number",
            "description": "Presión de diseño (psig)",
            "example": 50
          },
          "createdBy": {
            "description": "Usuario creador del tratamiento",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserResponseDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "name",
          "type",
          "totalFlow",
          "waterFraction",
          "inputTemperature",
          "treatmentTemperature",
          "ambientTemperature",
          "oilRetentionTime",
          "waterRetentionTime",
          "windSpeed",
          "apiGravity",
          "calculatedOilFlow",
          "calculatedWaterFlow",
          "oilRetentionVolume",
          "waterRetentionVolume",
          "requiredHeat",
          "heatLoss",
          "totalHeat",
          "selectedDiameter",
          "selectedLength",
          "designPressure",
          "createdBy"
        ]
      },
...
      "UpdateTreatmentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre del tratamiento",
            "example": "Tratamiento para crudo 18°API"
          },
          "description": {
            "type": "string",
            "description": "Descripción del tratamiento",
            "example": "Diseño para 500 bbl/día con 20% agua"
          },
          "type": {
            "type": "string",
            "description": "Tipo de tratador (vertical/horizontal)",
            "example": "horizontal",
            "enum": [
              "vertical",
              "horizontal"
            ]
          },
          "totalFlow": {
            "type": "number",
            "description": "Flujo total de emulsión (bpd)",
            "example": 500
          },
          "waterFraction": {
            "type": "number",
            "description": "Fracción de agua (%)",
            "example": 20
          },
          "inputTemperature": {
            "type": "number",
            "description": "Temperatura de entrada (°F)",
            "example": 75
          },
          "treatmentTemperature": {
            "type": "number",
            "description": "Temperatura de tratamiento (°F)",
            "example": 140
          },
          "ambientTemperature": {
            "type": "number",
            "description": "Temperatura ambiente (°F)",
            "example": 30
          },
          "oilRetentionTime": {
            "type": "number",
            "description": "Tiempo retención petróleo (min)",
            "example": 60
          },
          "waterRetentionTime": {
            "type": "number",
            "description": "Tiempo retención agua (min)",
            "example": 30
          },
          "windSpeed": {
            "type": "number",
            "description": "Velocidad del viento (mph)",
            "example": 15
          },
          "apiGravity": {
            "type": "number",
            "description": "Gravedad API del crudo",
            "example": 18
          }
        }
      },
      "CreateTreatmentOptionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Tipo de tratador (vertical/horizontal)",
            "example": "horizontal",
            "enum": [
              "vertical",
              "horizontal"
            ]
          },
          "diameter": {
            "type": "number",
            "description": "Diámetro exterior (ft)",
            "example": 4
          },
          "length": {
            "type": "number",
            "description": "Longitud estándar (ft)",
            "example": 15
          },
          "designPressure": {
            "type": "number",
            "description": "Presión de diseño (psig)",
            "example": 50
          },
          "minHeatCapacity": {
            "type": "number",
            "description": "Capacidad mínima de calor (BTU/hr)",
            "example": 250000
          },
          "notes": {
            "type": "string",
            "description": "Notas adicionales",
            "example": "Para crudos entre 15-25°API"
          }
        },
        "required": [
          "type",
          "diameter",
          "length",
          "designPressure",
          "minHeatCapacity"
        ]
      },
      "TreatmentOptionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Identificador único de la entidad.",
            "example": 1
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha y hora de creación del registro.",
            "example": "2023-10-27T10:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha y hora de la última actualización del registro.",
            "example": "2023-10-27T10:30:00Z"
          },
          "status": {
            "type": "boolean",
            "description": "Estado del registro (activo/inactivo).",
            "example": true
          },
          "deleted": {
            "type": "boolean",
            "description": "Indicador de borrado lógico.",
            "example": false
          },
          "type": {
            "type": "string",
            "description": "Tipo de tratador (vertical/horizontal)",
            "example": "horizontal",
            "enum": [
              "vertical",
              "horizontal"
            ]
          },
          "diameter": {
            "type": "number",
            "description": "Diámetro exterior (ft)",
            "example": 4
          },
          "length": {
            "type": "number",
            "description": "Longitud estándar (ft)",
            "example": 15
          },
          "designPressure": {
            "type": "number",
            "description": "Presión de diseño (psig)",
            "example": 50
          },
          "minHeatCapacity": {
            "type": "number",
            "description": "Capacidad mínima de calor (BTU/hr)",
            "example": 250000
          },
          "notes": {
            "type": "string",
            "description": "Notas adicionales",
            "example": "Para crudos entre 15-25°API"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "type",
          "diameter",
          "length",
          "designPressure",
          "minHeatCapacity"
        ]
      },
      "UpdateTreatmentOptionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Tipo de tratador (vertical/horizontal)",
            "example": "horizontal",
            "enum": [
              "vertical",
              "horizontal"
            ]
          },
          "diameter": {
            "type": "number",
            "description": "Diámetro exterior (ft)",
            "example": 4
          },
          "length": {
            "type": "number",
            "description": "Longitud estándar (ft)",
            "example": 15
          },
          "designPressure": {
            "type": "number",
            "description": "Presión de diseño (psig)",
            "example": 50
          },
          "minHeatCapacity": {
            "type": "number",
            "description": "Capacidad mínima de calor (BTU/hr)",
            "example": 250000
          },
          "notes": {
            "type": "string",
            "description": "Notas adicionales",
            "example": "Para crudos entre 15-25°API"
          }
        }
      },
...
    }
  }
}